+++ libraries/AP_BattMonitor/AP_BattMonitor_INA2xx.h
@@ -36,6 +36,7 @@ private:
         INA228,
         INA238,
         INA231,
+        INA219,
     };
 
     static const uint8_t i2c_probe_addresses[];