+++ libraries/AP_BattMonitor/AP_BattMonitor_Params.cpp
@@ -17,6 +17,13 @@
     #define AP_BATT_MONITOR_BATTERY_CAPACITY 3300
 #endif
 
+// Board-specific default battery monitor type (ESP32S3 uses INA219)
+#ifdef HAL_BATT_MONITOR_DEFAULT
+    #define AP_BATT_MONITOR_TYPE_DEFAULT HAL_BATT_MONITOR_DEFAULT
+#else
+    #define AP_BATT_MONITOR_TYPE_DEFAULT int8_t(AP_BattMonitor::Type::NONE)
+#endif
+
 const AP_Param::GroupInfo AP_BattMonitor_Params::var_info[] = {
     // @Param: MONITOR
     // @DisplayName: Battery monitoring
@@ -24,7 +31,7 @@ const AP_Param::GroupInfo AP_BattMonitor_Params::var_info[] = {
     // @Values: 0:Disabled,3:Analog Voltage Only,4:Analog Voltage and Current,5:Solo,6:Bebop,7:SMBus-Generic,8:DroneCAN-BatteryInfo,9:ESC,10:Sum Of Selected Monitors,11:FuelFlow,12:FuelLevelPWM,13:SMBUS-SUI3,14:SMBUS-SUI6,15:NeoDesign,16:SMBus-Maxell,17:Generator-Elec,18:Generator-Fuel,19:Rotoye,20:MPPT,21:INA2XX,22:LTC2946,23:Torqeedo,24:FuelLevelAnalog,25:Synthetic Current and Analog Voltage,26:INA239_SPI,27:EFI,28:AD7091R5,29:Scripting,30:INA3221
     // @User: Standard
     // @RebootRequired: True
-    AP_GROUPINFO_FLAGS("MONITOR", 1, AP_BattMonitor_Params, _type, int8_t(AP_BattMonitor::Type::NONE), AP_PARAM_FLAG_ENABLE),
+    AP_GROUPINFO_FLAGS("MONITOR", 1, AP_BattMonitor_Params, _type, AP_BATT_MONITOR_TYPE_DEFAULT, AP_PARAM_FLAG_ENABLE),
     
     // 2 was VOLT_PIN